Update value columns to be rational numbers

+++ db/migrate/20191002122545_create_adjustments.rb
@@ -4,7 +4,7 @@ class CreateAdjustments < ActiveRecord::Migration[6.0]
   def change
     create_table :adjustments do |t|
       t.text :reason, null: false
-      t.float :value, null: false
+      t.bigint :value, null: false
 
       t.integer :lock_version
       t.timestamps
